% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anovaDF.R
\name{anovaDF}
\alias{anovaDF}
\title{ANOVA overall P value for a linear model}
\usage{
anovaDF(vec, df)
}
\arguments{
\item{vec}{vector of the response variable (an eigen vector)}

\item{df}{data frame with the metadata}
}
\value{
a vector of P values for each model fitted where
  input vector is the response variable and dependent
  variable is on the of the columns in the provided
  data frame
}
\description{
The function calculates ANOVA overall P value for a linear
model where y is a numeric vector (an eigen vector
calculated in an SVD) and x is a dependent variable from
the metadata. The function loops over the columns of the
metadata
}
\seealso{
\code{\link{eigenVariable}}
}

