% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carefulKruskal.R
\name{carefulKruskal}
\alias{carefulKruskal}
\title{Calculate Kruskal-Wallis test between two variable, catch errors and warning}
\usage{
carefulKruskal(num.vec.ind, group.vec.ind, df)
}
\arguments{
\item{num.vec.ind}{name of a numeric variable in a data frame \code{df}}

\item{group.vec.ind}{name of a factor variable in a data frame \code{df}}

\item{df}{data frame}
}
\value{
\describe{
   \item{"p.value"}{"P value or the error/warning message"}
   \item{"statistic"}{"Test statistic or the error/warning message"}
   \item{"test.type"}{"String indicating the test type: kruskal-wallis rank sum test"}
 }
}
\description{
Given names of a numeric variable \code{x1} and \code{x2} in a data frame \code{df}
calculate the P value for the Kruskal-Wallis Rank Sum Test. Use \code{tryCatch}
to collect warnings and errors. Monte-Carlo simulation is performed by default
for more accurate P values.
}
\seealso{
\code{\link[stats]{kruskal.test}}
}

