% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carefulLM.R
\name{carefulLM}
\alias{carefulLM}
\title{Calculate P value for F test from the anova table, catch errors and warnings}
\usage{
carefulLM(resp.vec.ind, exp.vec.ind, df)
}
\arguments{
\item{resp.vec.ind}{index of the numeric response variable in a data frame \code{df}}

\item{exp.vec.ind}{index of the explanatory variable in a data frame \code{df}}

\item{df}{data frame}
}
\value{
\describe{
   \item{p.value}{P value or the error/warning message}
   \item{statistic}{Test statistic or the error/warning message}
   \item{test.type}{String indicating the test type: anova F test}
 }
}
\description{
Given names of a numeric \code{x1} and a factor \code{x2} variables in a data
frame \code{df} calculate the P value for the F test from the anova table
applied to linear regression.
Use \code{tryCatch} to collect warnings and errors.
}
\seealso{
\code{\link[stats]{anova}},
}

