% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carefulPearson.R
\name{carefulPearson}
\alias{carefulPearson}
\title{Pearson correlation test between two numeric variables, catch errors and warnings}
\usage{
carefulPearson(x1, x2, df, cor.test.method = "pearson")
}
\arguments{
\item{x1}{a character string indicating the name of a numeric variable in a data frame \code{df}}

\item{x2}{a character string indicating the name of another numeric variable in a data frame \code{df}}

\item{df}{data frame}

\item{cor.tes.method}{a character string indicating which correlation coefficient
is to be used for the test. Default is "pearson"}
}
\value{
\describe{
   \item{p.value}{P value or the error/warning message}
   \item{statistic}{Test statistic or the error/warning message}
   \item{test.type}{String indicating the test type: pearson correlation}
 }
}
\description{
Given names of two numeric variables \code{x1} and \code{x2} in a data frame \code{df}
calculate the P value for the Pearson's product moment correlation coefficient.
Use \code{tryCatch} to collect warnings and errors.
}
\seealso{
\code{\link[stats]{cor.test}}
}

