% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eigenVariable.R
\name{eigenVariable}
\alias{eigenVariable}
\title{Association of eigen vectors with variables}
\usage{
eigenVariable(pcs, tech.vars)
}
\arguments{
\item{pcs}{a matrix with eigen vectors, i.e V matrix of an SVD}

\item{tech.vars}{metadata df}
}
\value{
list with two elements. The first element is the matrix with the P values, and
 the second element is a ggplot object for the tile plot of the P values between
the variable, where color reflect the significance (alpha 0.05)
}
\description{
This function calculates P values for association of any number
 of eigen vectors with metadata variable.
 Method: anova(lm(eigengene ~ variable))$"Pr(>F)"[1]
}
\examples{
u <- matrix(runif(120), ncol = 3)
var.test <- data.frame(batch = sample(c("a", "b"), 40, replace = TRUE),
                       age = sample(50:85, size = 40, replace = TRUE),
                       amp.plate = sample(c("A1", "B1", "C1", "D1", "E1"), 40, replace = TRUE),
                       height = sample(60:72, 40, replace = TRUE))
res <- eigenVariable(u, var.test)
print(res$result)
plot(res$plot)
}
\seealso{
code{\link{anovaDF}}
}

