% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenVariablePlot.R
\name{eigenVariablePlot}
\alias{eigenVariablePlot}
\title{Plot an eigenvector vs a variable}
\usage{
eigenVariablePlot(pcs, variables, PC = 1, var.name, title = NULL)
}
\arguments{
\item{pcs}{a matrix with eigen vectors, V matrix of an SVD}

\item{variables}{a data frame with clinical/technical variables pertaining to the data}

\item{PC}{number of PC to plot, default is 1 (first)}

\item{var.name}{name of the variable from \code{variable} data frame}

\item{title}{plot title}
}
\value{
ggplot object
}
\description{
The function select an approproate plot to show a relationship between
an eigenvector and a variable from the metadata. If the variable is numeric
or integer a scatterplot will be produced, if the variable is a factor or
a character a boxplot will be produced. The eigenvector will be plotted on
the y axis a and the variable from the metadata will be plotted on the x
axis. The plot is based on ggplot.
}
\seealso{
code{\link{fs}}
}

