% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fs.R
\name{fs}
\alias{fs}
\title{Calculate an SVD on a matrix}
\usage{
fs(x)
}
\arguments{
\item{x}{data matrix}
}
\value{
a list with U, V and D matrices. Column names of V get names: EigenN,
 where N is the number of the eigenvector. Row names of V get the column names
 of the input data matrix. Column names of U get the column names of the input
 data matrix \code{x}.
}
\description{
Use \code{\link[corpcor]{fast.svd}} from \code{corpcor} library to calculate an SVD
 of a matrix. The data before the calculation is centered. The D
 matrix is transformed to show percent variance explained.
}
\details{
The function is a slight modification of the function created by Brig Mecham.
}
\references{
\href{https://github.com/Sage-Bionetworks/mGenomics/blob/master/R/fs.R}{mGenomics}
}
\seealso{
\code{\link[corpcor]{fast.svd}}
}

