% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOntology.R
\name{gostats}
\alias{gostats}
\title{Obtain Gene Ontology (GO) results for a gene list}
\usage{
gostats(data, background, annotation)
}
\arguments{
\item{data}{a vector of ENTREZ Gene IDs.}

\item{background}{a vector ENTREZ Gene IDs used as the universe
for the hypergeometric calculation, also ENTREZ Gene IDs.}

\item{annotation}{- a string giving the name of the annotation data package with
GO annotation. See \code{\link[GOstats]{GOHyperGParams-class}}.}
}
\value{
data frame object with GO results. Columns: Ontology (MF, BP or CC),
GOID, Pvalue, OddsRatio, ExpCount, Count, Size, Term, P.Bonferroni
}
\description{
This function performs Gene Ontology analysis on a list of gene. It obtains
MF, BP and CC categories for a gene list, calculates enrichment P value and
multiple testing correction based on Bonferroni method. Finally for each MF,
BP and CC it selects only the categories for which corrected P value is less
or equal to 0.05. Returns a data frame with GO categories.
}
\seealso{
\code{\link{fs}}
}

