% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixOutliers.R
\name{matrixOutliers}
\alias{matrixOutliers}
\title{Plots data outliers based on first eigen vector}
\usage{
matrixOutliers(data, pc)
}
\arguments{
\item{data}{a matrix for which an SVD was computed}

\item{pc}{vector corresponding to an eigen vector of the SVD}
}
\value{
a ggplot object
}
\description{
The function takes the eigen vector provided, orders it's values,
takes the minimum and the maximum, finds the corresponding columns
in the provided data matrix and creates a scatter plot of these
vectors. Adds a line with intercept 0 and the slope 1. This kind of
plot is generally produced for the first eigen vector.
}
\seealso{
\code{\link{fs}}
}

