% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcScatter.R
\name{pcScatter}
\alias{pcScatter}
\title{Plot a scatter plot between two eigenvectors colored by a variable}
\usage{
pcScatter(eigens, metadata, pc.to.plot, variable, title = NULL)
}
\arguments{
\item{eigens}{a matrix with eigenvectors}

\item{metadata}{a data frame with the variable, row names and column names
must be present}

\item{pc.to.plot}{a vector with two elements indicating for which two eigenvector
the plot should be created}

\item{variable}{by which the points of the scatter plot should be colored}
}
\value{
plot and the merged pc/metadata data frame
}
\description{
This function combines eigenvectors with the
metadata and creates a scatter plot between 2 eigenvectors and colors
them by the third variable present in the metadata. Matching of the
metadata and the PCs is also executed
}
\seealso{
fs
}

