% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotVariance.R
\name{plotVariance}
\alias{plotVariance}
\title{Plot percent variance explained}
\usage{
plotVariance(u, plot.title = NULL)
}
\arguments{
\item{u}{list with the SVD components: u, v, d matrices.
It assumes that d is already transformed to reflect the percent
variance. Use fs function to generate the SVD}

\item{plot.title}{plot title}
}
\value{
ggplot object
}
\description{
Based on ggplot, prettier than the basic graphics
}
\seealso{
fs
}

