% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryTable.R
\name{summaryTable}
\alias{summaryTable}
\title{Summarize each column of a data frame}
\usage{
summaryTable(df)
}
\arguments{
\item{df}{a data frame for columns of which the summary statistics are to
be calculated}
}
\value{
a data frame with seven columns: type of the variable, number of
categories (when a factor or a character), minimum, mean, median, maximum,
number of the missing cases. Row names
of the data frame are the column names of the input data frame.
}
\description{
Function takes a data frame and creates a series of summary statistics for each
column of the data frame based on its class. If a column is numeric or integer,
the function provides the minimum,
median, mean, maximum, the number of the missing values and the class.
If the class is a factor it provides the number of levels, and the statistics for
the number of cases as defined by each level: the minimum,
the median, the mean, the maximum and the number of missing cases. If the class
is a character it is then converted to a factor and the same
statistics are given.
}

