% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testDecisionTree.R
\name{testDecisionTree}
\alias{testDecisionTree}
\title{Decide what correlation test to perform and obtain statistics}
\usage{
testDecisionTree(ind1, ind2, df, chisq.p.val.sim = TRUE,
  cor.method = "pearson")
}
\arguments{
\item{ind1}{a string to indicate the name of the first variable}

\item{ind2}{a string to indicate the name of the second variable}

\item{df}{a data frame with the variables}

\item{chisq.p.val.sim}{boolean to indicate whether the simulation should
be performed for the Chi-square test. Default is TRUE}

\item{cor.method}{string to indicate what method should be used to compute
correlation between two numeric variables}
}
\value{
\describe{
   \item{p.value}{P value for the test}
   \item{statistic}{Test statistic}
   \item{test.type}{Type of the test performed: Chi-square, Pearson or Kruskal-Wallis}
 }
}
\description{
Given the names of two variable \code{ind1} and \code{ind2} in a data frame
\code{df}
decide which test should be performed to calculate the relationship
between these variables. The decision is based on their respective classes.
Class pairs and the corresponing
tests: factor/factor - Chi-square (with or without a Monte-Carlo simulation),
numeric/numeric - Pearson correlation, numeric/factor - Kruskal-Wallis test.
}
\seealso{
\code{\link{variablesRelation}}, \code{\link{carefulChisq}}, \code{\link{carefulKruskal}},
\code{\link{carefulPearson}}
}

