% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variableSummary.R
\name{variableSummary}
\alias{variableSummary}
\title{Single variable summary for the metadata}
\usage{
variableSummary(data.var)
}
\arguments{
\item{data.var}{a vector of integer, numeric, character of factor class}
}
\value{
a named vector with seven elements: type of the variable, number of
categories, minimum, mean, median, maximum, number of missing values
}
\description{
This function takes a variable and creates a series of summary statistics based
on its class. If the vector is numeric or integer, the function provides the minimum,
median, mean, maximum, number of missing values and the type of the variable.
If the class is factor it provides the number of levels, and the statistics for
the number of cases as defined by each level: the minimum,
the median, the mean, the maximum and the number of missing cases. If the class
is character then it is converted to a factor and the same statistics as for the
factor are given.
}
\seealso{
\code{\link{summaryTable}}
}

