% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carefulChisq.R
\name{carefulChisq}
\alias{carefulChisq}
\title{Calculate Chi-square test between two variables, catch errors and warnings}
\usage{
carefulChisq(x1, x2, df, sim.p.value = TRUE)
}
\arguments{
\item{x1}{name of one factor variable in a data frame \code{df}}

\item{x2}{name of another factor variable in a data frame \code{df}}

\item{df}{data frame}

\item{sim.p.value}{if it TRUE (default) P value is computed for a Monte carlo test}
}
\value{
\describe{
   \item{p.value}{P value or the error/warning message}
   \item{statistic}{Test statistic or the error/warning message}
   \item{test.type}{String indicating the test type: chi-square test,
   monte carlo simulation}
 }
}
\description{
Given names of two variables \code{x1} and \code{x2} in a data frame \code{df}
calculate the P value for the Chi-square test between them. Use \code{tryCatch}
to collect warnings and errors. Monte-Carlo simulation is performed by default
for more accurate P values.
}
\seealso{
\code{\link[stats]{chisq.test}}
}

