% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixOutliers.R
\name{matrixOutliers}
\alias{matrixOutliers}
\title{Plots data outliers based on first eigenvector}
\usage{
matrixOutliers(data, pc)
}
\arguments{
\item{data}{dataset for which an SVD was computed}

\item{pc}{vector corresponding to an eigenvector of the SVD}
}
\value{
a ggplot object
}
\description{
The function takes the eigenvector provided, orders it's values,
takes the minimum and the maximum, finds the corresponding columns
in the provided data matrix and creates a scatter plot of these
columns. Adds a line with intercept 0 and the slope 1.
}
\seealso{
fs
}

