% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcScatter.R
\name{pcScatter}
\alias{pcScatter}
\title{Create a scatter plot between two eigen vectors colored by a variable}
\usage{
pcScatter(eigens, metadata, pc.to.plot, variable, title = NULL)
}
\arguments{
\item{eigens}{a matrix with the eigen vectors}

\item{metadata}{a data frame with the variable of interest (points color), must have row and column names}

\item{pc.to.plot}{a numeric vector with two elements indicating which two
eigen vectors should be used for a scatter plot}

\item{variable}{a string indicating a name of the variable in
the metadata by which the points of the scatter plot should be colored}

\item{title}{string for the plot title. Default is NULL - no plot title}
}
\value{
\describe{
   \item{plot}{ggplot2 plot object}
   \item{eigen.meta}{data frame of the metadata merged with the matrix of the eigen vectors}
 }
}
\description{
This function combines eigen vectors with the
metadata, creates a scatter plot (ggplot2 graphics) between two eigenvectors and colors
the points by a variable present in the metadata. The metadata and the
matrix of the eigen vectors are matched by row names.
}
\seealso{
\code{\link{fs}}
}

