% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariance.R
\name{plotVariance}
\alias{plotVariance}
\title{Plot percent variance explained}
\usage{
plotVariance(d, plot.title = NULL)
}
\arguments{
\item{d}{a vector containing the values for the variance explained by each
singular vector}

\item{plot.title}{plot title}
}
\value{
ggplot object
}
\description{
Based on ggplot2 graphics library, prettier than the basic graphics
}
\seealso{
\code{\link{fs}}
}

