% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testDecisionTree.R
\name{testDecisionTree}
\alias{testDecisionTree}
\title{Decide what correlation test to perform and obtain statistics}
\usage{
testDecisionTree(ind1, ind2, df, chisq.p.val.sim = TRUE,
  cor.method = "pearson")
}
\value{
\describe{
   \item{"p.value"}{"P value for the test"}
   \item{"statistic"}{"Test statistic"}
   \item{"test.type"}{"Type of the test performed: Chi-Square, Pearson or Kruskal-Wallis"}
 }
}
\description{
Given the names of two variable \code{ind1} and \code{ind2} in a data frame
\code{df}
decide which test should be performed to calculate the relationship
between the variables based on their classes. Class pairs and the corresponing
tests: factor/factor - Chi-square (with or without a Monte-Carlo simulation).
numeric/numeric - Pearson correlation, numeric/factor - Kruskal-Wallis test.
}
\seealso{
\code{\link{variablesRelation}}, \code{\link{carefulChisq}}, \code{\link{carefulKruskal}},
\code{\link{carefulPearson}}
}

