% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variablesRelation.R
\name{variablesRelation}
\alias{variablesRelation}
\title{Calculate all pairwise variable relations}
\usage{
variablesRelation(df, multiple.testing = c("fdr", "BY", "q"))
}
\arguments{
\item{df}{data frame for which relationship between columns needs to be calculated}

\item{multiple.testing}{method to use for multiple testing adjustment, currently
supported methods are Benjamini & Hochberg's FDR ("fdr"), Benjamini & Yekutieli ("BY")
or Storey's q-value ("q")}
}
\value{
\describe{
   \item{plot}{ggplot object with the tile plot where two colors are used to
   highlight significant relationships (alpha is less or equal to 0.05)}
   \item{table}{a data frame with all pairwise tests: variable 1 name,
   variable 2 name, P value, test statistic, test name, adjusted P value using
   one of methods}
 }
}
\description{
Given a data frame the function will calculate pairwise correlations
between all columns. This function is usually used to understand if any of the
technical variables in the data are correlated with the biological variables
(to avoid loosing the signal during data normalization).
The following tests are used: between a factor and a numeric
variable or between two numeric variables - anova F test,
between two factor variables - Chi-square test. If simulate.p.val = TRUE then
the p-value for the Chi-square test is computed for a Monte Carlo test with
monte.carlo.reps replicates. We suggest removing any columns ID columns from the
data prior to using this function (definition of an ID column: number of levels
is the same as the number of rows in the data frame)
}

